Moved App component into separate file to reflect changes made in main and navigation branches. Probably not the appropriate practice.
